Move efficiency components to sub-directories
