+++ apps/koku-ui-hccm/src/routes/optimizations/efficiency/components/compute/computeCard.scss
