+++ apps/koku-ui-hccm/src/routes/optimizations/efficiency/components/memory/index.ts
