+++ apps/koku-ui-hccm/src/routes/optimizations/efficiency/components/summary/efficiencySummary.scss
