+++ apps/koku-ui-hccm/src/routes/optimizations/efficiency/efficiency.styles.ts
@@ -1,5 +1,4 @@
 import t_global_spacer_lg from '@patternfly/react-tokens/dist/js/t_global_spacer_lg';
-import t_global_spacer_sm from '@patternfly/react-tokens/dist/js/t_global_spacer_sm';
 import type React from 'react';
 
 export const styles = {
@@ -8,15 +7,5 @@ export const styles = {
   },
   headerContainer: {
     paddingBottom: 0,
-  },
-  export: {
-    marginTop: '-5px',
-  },
-  paginationContainer: {
-    marginTop: t_global_spacer_sm.var,
-  },
-  tableTitle: {
-    display: 'flex',
-    marginTop: t_global_spacer_sm.var,
   },
 } as { [className: string]: React.CSSProperties };
